requirements.txt
alembic==1.8.0
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.0.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
Pillow==9.1.1
psycopg2-binary==2.9.3
py3dns==3.2.1
pycparser==2.21
python-dotenv==0.20.0
requests==2.28.0
SQLAlchemy==1.4.37
stripe==3.4.0
urllib3==1.26.9
validate-email-address==1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
